Garbage Collection
This is the cleanup part for the queue
handling of add --queuemode / processqueue,
because for safety reasons the processed entries
are not yet deleted there yet.

This could either become a no-op once the queue
mode is death-proof (but who wants to challenge
Stuntman Mike), or it can take more of garbage
collection in the future.